backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
httpx
python-multipart
python-dotenv
pdfplumber
python-docx
passlib
bcrypt
groq
pydantic
hindsight-client
